requirements.txt
alabaster==0.7.8
alembic==1.7.1
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==21.1.0
asgiref==3.4.1
async-generator==1.10
atomicwrites==1.1.5
attrs==21.2.0
Automat==0.8.0
Babel==2.6.0
backcall==0.1.0
bcrypt==3.1.7
bleach==4.1.0
blinker==1.4
brial==1.0.2
Brlapi==0.7.0
certifi==2019.11.28
certipy==0.1.3
cffi==1.14.6
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
cupshelpers==1.0
cvxopt==1.2.3
cycler==0.10.0
cypari2==2.1.1
Cython==0.29.14
dbus-python==1.2.16
debugpy==1.4.1
decorator==4.4.2
decouple==0.0.7
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.7
django-cors-headers==3.10.0
django-crispy-forms==1.6.0
django-filter==21.1
django-jazzmin==2.4.8
django-markdown-deux==1.0.5
django-pagedown==0.1.1
django-taggit==1.5.1
django-templated-mail==1.1.1
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
fpylll==0.5.1.dev0
future==0.18.2
gmpy2==2.1.0b3
greenlet==1.1.1
gunicorn==20.1.0
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
importlib-resources==5.2.2
incremental==16.10.1
ipykernel==6.3.1
ipython==7.27.0
ipython_genutils==0.2.0
ipywidgets==7.6.4
itsdangerous==1.1.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.2
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-telemetry==0.1.0
jupyterhub==1.4.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.1
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
markdown2==2.3.1
MarkupSafe==2.0.1
matplotlib==3.1.2
matplotlib-inline==0.1.2
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
mpmath==1.1.0
mysql-connector==2.2.9
mysql-connector-python==8.0.26
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
netifaces==0.10.4
networkx==2.4
nose==1.3.7
notebook==6.4.3
numpy==1.21.2
oauthlib==3.1.0
olefile==0.46
packaging==21.0
pamela==1.0.0
pandas==1.3.3
pandocfilters==1.4.3
paramiko==2.6.0
parso==0.8.2
path.py==12.0.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==3.1.0
pkgconfig==1.5.1
pluggy==0.13.0
pplpy==0.8.4
prometheus-client==0.11.0
prompt-toolkit==2.0.10
protobuf==3.6.1
psutil==5.5.1
ptyprocess==0.7.0
py==1.8.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.20
pycups==1.9.73
Pygments==2.10.0
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==2.1.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.0
PyStemmer==1.3.0
pytest==4.6.9
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-json-logger==2.0.2
python3-openid==3.1.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.11.0
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
roman==2.0.0
rpy2==3.2.6
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sage==9.0
sagenb-export==3.2
scipy==1.3.3
SecretStorage==2.3.1
Send2Trash==1.8.0
service-identity==18.1.0
setuptools-scm==3.4.3
simplegeneric==0.8.1
simplejson==3.16.0
six==1.14.0
snowballstemmer==2.0.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
Sphinx==1.8.5
SQLAlchemy==1.4.23
sqlparse==0.4.1
sympy==1.5.1
systemd-python==234
terminado==0.11.1
testpath==0.5.0
toml==0.10.0
tornado==6.1
traitlets==5.1.0
Twisted==18.9.0
tzlocal==2.1b1
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.8
usb-creator==0.3.7
vcversioner==2.16.0.0
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.1
whitenoise==5.3.0
widgetsnbextension==3.5.1
xkit==0.0.0
zipp==3.5.0
zope.interface==4.7.1
